course_management_ml/requirements.txt
catboost==1.2.8
colorlog==6.9.0
cycler==0.12.1
fonttools==4.60.1
graphviz==0.21
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
mysql==0.0.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
narwhals==2.10.0
plotly==6.3.1
pyparsing==3.2.5
pyqt6-webengine==6.9.0
pyqt6-webengine-qt6==6.9.2
pytz==2025.2
scienceplots==2.1.1
scikit-learn==1.6.1
scipy==1.16.2
seaborn==0.13.2
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.1.1